Implement on ResumesController

+++ database/migrations/2026_02_12_125656_create_company_socials_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('company_socials', function (Blueprint $table) {
             $table->id();
-            $table->integer('company_id');
+            $table->unsignedBigInteger('company_id');
             $table->string('platform')->nullable(); // Facebook or Telegram. etc.
             $table->text('url')->nullable();
             $table->timestamps();
