+++ database/migrations/2026_02_12_125707_create_companies_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('companies', function (Blueprint $table) {
             $table->id();
-            $table->integer("user_id");
+            $table->unsignedBigInteger("user_id");
             $table->string("name");
             $table->string("slug")->nullable();
             $table->string("logo_path")->nullable();
