+++ database/migrations/2026_02_12_125900_create_save_jobs_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('save_jobs', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('job_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('job_id');
             $table->timestamps();
         });
     }
